requirements.txt
wxPython~=4.0.0a1
six~=1.16.0
numpy~=1.16.6
mysql-connector-python~=8.0.5
pandas~=0.24.2
setuptools~=44.1.1
future~=0.18.2
pytz~=2022.1
wheel~=0.37.1
protobuf~=3.17.3
python-dateutil~=2.8.2
altgraph~=0.17.2
pefile~=2019.4.18
dis3~=0.1.3
PyInstaller~=3.6